Revert "Citizen UI add function tests"
