+++ package.json
@@ -17,11 +17,10 @@
     "test:a11y": "jest -c jest.a11y.config.js --maxWorkers 15 --forceExit",
     "sonar-scan": "sonar-scanner",
     "test:smoke": "jest -c jest.smoketest.config.js",
+    "test:functional": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs --config ./src/test/functional/codecept.conf.js run-workers 2 parallel",
     "test:crossbrowser": "SAUCE='true' yarn test:crossbrowser:codeceptjs run",
-    "test:functional": "yarn test:functional:codeceptjs run --grep @test --reporter mocha-multi --verbose",
     "test:crossbrowser:playwright": "yarn test:crossbrowser:codeceptjs run-multiple crossBrowser",
     "test:crossbrowser:codeceptjs": "codeceptjs --config ./src/test/cross-browser/codecept.conf.js",
-    "test:functional:codeceptjs": "codeceptjs --config src/test/functionalTests/codecept.conf.js",
     "cichecks": "yarn && yarn build && yarn lint  && yarn test && yarn test:routes &&  yarn test:a11y"
   },
   "dependencies": {
@@ -72,8 +71,6 @@
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.20",
     "luxon": "^2.4.0",
-    "mocha-multi": "^1.1.6",
-    "mochawesome": "^7.1.3",
     "nunjucks": "^3.2.2",
     "otplib": "^12.0.1",
     "reflect-metadata": "^0.1.13",
@@ -95,7 +92,7 @@
     "@typescript-eslint/eslint-plugin-tslint": "^5.30.4",
     "@typescript-eslint/parser": "^5.30.4",
     "babel-loader": "^8.2.5",
-    "codeceptjs": "^3.3.3",
+    "codeceptjs": "^3.3.0",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
     "debug": "^4.3.4",
@@ -107,8 +104,6 @@
     "mini-css-extract-plugin": "^2.6.1",
     "nock": "^13.2.8",
     "nodemon": "^2.0.18",
-    "mocha": "^10.0.0",
-    "mocha-junit-reporter": "^2.0.2",
     "pa11y": "^6.2.3",
     "playwright": "^1.23.1",
     "request": "^2.88.0",
