+++ src/test/e2e/step_definitions/steps.ts
@@ -1,0 +1,3 @@
+const { phonenumber } = require('../pages/phonenumber');
+const { I } = inject();
+phonenumber(I);
