--- src/test/functionalTests/features/dashboard/pages/dashboard.ts
@@ -1,13 +1,0 @@
-import I = CodeceptJS.I
-
-const I: I = actor();
-
-export class DashboardPage {
-
-  open (): void {
-    I.amOnPage('/dashboard');
-  }
-  verifyDashboardPageContent (): void {
-    I.see('Claims made against you');
-  }
-}
