--- src/test/functionalTests/features/response/pages/explanation.ts
@@ -1,18 +1,0 @@
-import I = CodeceptJS.I
-
-const I: I = actor();
-
-const fields ={
-  reason: 'textarea[id="text"]',
-};
-const buttons = {
-  continue: 'button.govuk-button',
-};
-
-export class Explanation {
-
-  enterExplanation(): void {
-    I.fillField(fields.reason, 'Test reason');
-    I.click(buttons.continue);
-  }
-}
