--- src/test/functionalTests/features/response/pages/selfEmploymentTaxDetails.ts
@@ -1,29 +1,0 @@
-import I = CodeceptJS.I
-
-const I: I = actor();
-
-const fields ={
-  yesButton: 'input[id="option"]',
-  noButton: 'input[id="option-2"]',
-  amountYouOwe: 'input[id="amountYouOwe"]',
-  reason: 'textarea[id="reason"]',
-};
-const buttons = {
-  continue: 'button.govuk-button',
-};
-
-export class SelfEmploymentTaxDetails {
-
-  clickYesButton(): void {
-    I.see('Are you behind on tax payments?', 'h1');
-    I.click(fields.yesButton);
-    I.fillField(fields.amountYouOwe, '2000');
-    I.fillField(fields.reason, 'Last year pending');
-    I.click(buttons.continue);
-  }
-  clickNoButton(): void {
-    I.see('Are you behind on tax payments?', 'h1');
-    I.click(fields.noButton);
-    I.click(buttons.continue);
-  }
-}
