--- src/test/functionalTests/features/response/pages/taskList.ts
@@ -1,14 +1,0 @@
-import I = CodeceptJS.I
-
-const I: I = actor();
-
-export class TaskListPage {
-
-  open(claimRef): void {
-    I.amOnPage('/case/' + claimRef + '/response/task-list');
-  }
-
-  verifyResponsePageContent(): void {
-    I.see('Respond to a money claim');
-  }
-}
